relocate lists restore collection test
